requirements.txt
streamlit
spacy
torch
transformers
langchain-neo4j
langchain-ollama
numpy
scikit-fuzzy
python-dotenv
streamlit-authenticator